+++ packages/components/src/Dialog/Dialog.tsx
@@ -1,14 +1,46 @@
-import { useContext } from 'react';
+import {
+  ForwardRefExoticComponent,
+  ReactNode,
+  Ref,
+  RefAttributes,
+  forwardRef,
+  useContext,
+} from 'react';
 import type RAC from 'react-aria-components';
 import { Dialog, OverlayTriggerStateContext } from 'react-aria-components';
 import { cn, useClassNames } from '@marigold/system';
 import { DialogActions } from './DialogActions';
 import { DialogContent } from './DialogContent';
 import { DialogTitle } from './DialogTitle';
-import { DialogTrigger } from './DialogTrigger';
+import { DialogContext, DialogTrigger } from './DialogTrigger';
+import { Modal } from './Modal';
 
-// Close Button
-// ---------------
+export interface DialogProps
+  extends Omit<RAC.DialogProps, 'className' | 'style'> {
+  variant?: string;
+  size?: string;
+  /**
+   * Show the close button.
+   */
+  closeButton?: boolean;
+}
+
+interface DialogComponent
+  extends ForwardRefExoticComponent<
+    DialogProps & RefAttributes<HTMLInputElement>
+  > {
+  /**
+   * Options for the Combobox.
+   */
+  Trigger: typeof DialogTrigger;
+
+  Title: typeof DialogTitle;
+
+  Content: typeof DialogContent;
+
+  Actions: typeof DialogActions;
+}
+
 interface CloseButtonProps {
   className?: string;
 }
@@ -35,56 +67,39 @@ const CloseButton = ({ className }: CloseButtonProps) => {
   );
 };
 
-// Props
-// ---------------
-export interface DialogProps
-  extends Omit<RAC.DialogProps, 'className' | 'style'> {
-  variant?: string;
-  size?: string;
-  /**
-   * Show the close button.
-   */
-  closeButton?: boolean;
-  /**
-   * If `true`, the dialog will be non-modal, meaning it will not block interaction with the background content.
-   * @default false
-   */
-  isNonModal?: boolean;
-}
-
-// Component
-// ---------------
-const _Dialog = ({
-  variant,
-  size,
-  closeButton,
-  isNonModal,
-  ...props
-}: DialogProps) => {
-  const classNames = useClassNames({ component: 'Dialog', variant, size });
-  const state = useContext(OverlayTriggerStateContext);
-
-  const children =
-    typeof props.children === 'function'
-      ? props.children({
-          close: state?.close ?? (() => {}),
-        })
-      : props.children;
-
-  return (
-    <Dialog
-      {...props}
-      className={cn(
-        'relative outline-hidden [&>*:not(:last-child)]:mb-4',
-        "grid [grid-template-areas:'title'_'content'_'actions']",
-        classNames.container
-      )}
-    >
-      {closeButton && <CloseButton className={classNames.closeButton} />}
-      {children}
-    </Dialog>
-  );
-};
+const _Dialog = forwardRef(
+  (props: DialogProps, ref: Ref<HTMLElement> | undefined) => {
+    const classNames = useClassNames({
+      component: 'Dialog',
+      variant: props.variant,
+      size: props.size,
+    });
+    const { isDismissable, isKeyboardDismissDisabled, isOpen } =
+      useContext(DialogContext);
+    return (
+      <Modal
+        isDismissable={isDismissable}
+        isKeyboardDismissDisabled={isKeyboardDismissDisabled}
+        isOpen={isOpen}
+      >
+        <Dialog
+          {...props}
+          ref={ref}
+          className={cn(
+            'relative outline-hidden [&>*:not(:last-child)]:mb-4',
+            "grid [grid-template-areas:'title'_'content'_'actions']",
+            classNames.container
+          )}
+        >
+          {props.closeButton && (
+            <CloseButton className={classNames.closeButton} />
+          )}
+          {props.children as ReactNode}
+        </Dialog>
+      </Modal>
+    );
+  }
+) as DialogComponent;
 
 _Dialog.Trigger = DialogTrigger;
 _Dialog.Title = DialogTitle;
