+++ packages/components/src/Overlay/Overlay.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { forwardRef } from 'react';
 import { Menu, MenuItem, MenuTrigger } from 'react-aria-components';
 import { Button } from '../Button';
-import { Dialog } from '../Dialog';
+import { Dialog } from '../Dialog_dep';
 import { Headline } from '../Headline';
 import { Text } from '../Text';
 import { Modal } from './Modal';
